+++ src/text/font_loader.rs
@@ -1,12 +1,32 @@
-use super::font::VelloFont;
-use crate::integrations::VectorLoaderError;
+use super::{context::get_global_font_context, font::VelloFont};
+use crate::{integrations::VectorLoaderError, text::context::LOCAL_FONT_CONTEXT};
 use bevy::{
     asset::{AssetLoader, LoadContext, io::Reader},
-    utils::ConditionalSendFuture,
+    log::warn,
 };
 
 #[derive(Default)]
 pub struct VelloFontLoader;
+
+pub(crate) fn load_into_font_context(bytes: Vec<u8>) -> VelloFont {
+    LOCAL_FONT_CONTEXT.with_borrow_mut(|font_context| {
+        if font_context.is_none() {
+            *font_context = Some(get_global_font_context().clone());
+        }
+        let font_context = font_context.as_mut().unwrap();
+        let registered_fonts = font_context.collection.register_fonts(bytes.clone());
+        let maybe_font = registered_fonts.first();
+        if maybe_font.is_none() {
+            warn!("Failed to register default font");
+        }
+        let (family_id, _font_info_vec) = maybe_font.unwrap();
+        let family_name = font_context.collection.family_name(*family_id).unwrap();
+        VelloFont {
+            family_name: family_name.to_string(),
+            bytes,
+        }
+    })
+}
 
 impl AssetLoader for VelloFontLoader {
     type Asset = VelloFont;
@@ -15,19 +35,15 @@ impl AssetLoader for VelloFontLoader {
 
     type Error = VectorLoaderError;
 
-    fn load(
+    async fn load(
         &self,
         reader: &mut dyn Reader,
         _settings: &Self::Settings,
-        _load_context: &mut LoadContext,
-    ) -> impl ConditionalSendFuture<Output = Result<Self::Asset, Self::Error>> {
-        Box::pin(async move {
-            let mut bytes = Vec::new();
-            reader.read_to_end(&mut bytes).await?;
-            let vello_font = VelloFont::new(bytes.to_vec());
-
-            Ok(vello_font)
-        })
+        _load_context: &mut LoadContext<'_>,
+    ) -> Result<Self::Asset, Self::Error> {
+        let mut bytes = Vec::new();
+        reader.read_to_end(&mut bytes).await?;
+        Ok(load_into_font_context(bytes))
     }
 
     fn extensions(&self) -> &[&str] {