+++ src/text/vello_text.rs
@@ -9,21 +9,109 @@ pub struct VelloTextSection {
     pub style: VelloTextStyle,
 }
 
-#[derive(Component, Clone)]
+#[derive(Clone)]
 pub struct VelloTextStyle {
     pub font: Handle<VelloFont>,
-    pub font_size: f32,
     pub brush: Brush,
+    pub font_size: f32,
+    /// Line height multiplier.
+    pub line_height: f32,
+    /// Extra spacing between words.
+    pub word_spacing: f32,
+    /// Extra spacing between letters.
+    pub letter_spacing: f32,
+    pub font_axes: VelloFontAxes,
 }
 
 impl Default for VelloTextStyle {
     fn default() -> Self {
         Self {
             font: Default::default(),
-            font_size: 24.0,
             brush: Brush::Solid(peniko::Color::WHITE),
+            font_size: 24.0,
+            line_height: 1.0,
+            word_spacing: 0.0,
+            letter_spacing: 0.0,
+            font_axes: Default::default(),
         }
     }
+}
+
+/// Describes the variable axes of a font.
+///
+/// https://fonts.google.com/knowledge/introducing_type/introducing_variable_fonts
+///
+/// Each axis is optional and only present if the font supports it.
+#[derive(Default, Clone)]
+pub struct VelloFontAxes {
+    /// wght variable axis only if the font supports it
+    ///
+    /// https://fonts.google.com/knowledge/glossary/weight_axis
+    pub weight: Option<f32>,
+    /// wdth variable axis only if the font supports it
+    ///
+    /// https://fonts.google.com/knowledge/glossary/width_axis
+    pub width: Option<f32>,
+    /// opsz variable axis only if the font supports it
+    ///
+    /// https://fonts.google.com/knowledge/glossary/optical_size_axis
+    pub optical_size: Option<f32>,
+    /// ital variable axis only if the font supports it
+    ///
+    /// Mutually exclusive with `slant`.
+    ///
+    /// https://fonts.google.com/knowledge/glossary/italic_axis
+    pub italic: bool,
+    /// slnt variable axis only if the font supports it
+    ///
+    /// Mutually exclusive with `italic`.
+    ///
+    /// If italic is true, slant will be ignored.
+    ///
+    /// https://fonts.google.com/knowledge/glossary/slant_axis
+    pub slant: Option<f32>,
+
+    /// GRAD variable axis only if the font supports it
+    ///
+    /// https://fonts.google.com/knowledge/glossary/grade_axis
+    pub grade: Option<f32>,
+
+    /// XOPQ variable axis only if the font supports it
+    ///
+    /// https://fonts.google.com/knowledge/glossary/xopq_axis
+    pub thick_stroke: Option<f32>,
+    /// yopq variable axis only if the font supports it
+    ///
+    /// https://fonts.google.com/knowledge/glossary/yopq_axis
+    pub thin_stroke: Option<f32>,
+
+    /// XTRA variable axis only if the font supports it
+    ///
+    /// https://fonts.google.com/knowledge/glossary/xtra_axis
+    pub counter_width: Option<f32>,
+
+    /// YTUC variable axis only if the font supports it
+    ///
+    /// https://fonts.google.com/knowledge/glossary/ytuc_axis
+    pub uppercase_height: Option<f32>,
+    /// YTLC variable axis only if the font supports it
+    ///
+    /// https://fonts.google.com/knowledge/glossary/ytlc_axis
+    pub lowercase_height: Option<f32>,
+
+    /// YTAS variable axis only if the font supports it
+    ///
+    /// https://fonts.google.com/knowledge/glossary/ytas_axis
+    pub ascender_height: Option<f32>,
+    /// YTDE variable axis only if the font supports it
+    ///
+    /// https://fonts.google.com/knowledge/glossary/ytde_axis
+    pub descender_depth: Option<f32>,
+
+    /// YTFI variable axis only if the font supports it
+    ///
+    /// https://fonts.google.com/knowledge/glossary/ytfi_axis
+    pub figure_height: Option<f32>,
 }
 
 /// Describes how the text is positioned relative to its [`Transform`]. It defaults to [`VelloTextAnchor::BottomLeft`].